gopls/internal/golang: fix DocLink markup
This CL fixes a bug in the DocLink markup used by
"Browse documentation for F" that arose from using
astutil.NodeContains on an ast.File: the Pos/End range
of a file is not the complete file, and excludes the
package doc comment, among other things.

Also, support doclinks [pkg.Sym] where pkg is not imported
from the file containing the doc comment, but is imported
by some other file in the package.

+ Test

We really need a way to write unit tests of this logic since
it could use a lot more coverage, and integration tests of
RPCs are a cumbersome and inefficient way to test text markup.

Fixes golang/go#70905

Change-Id: Ib3994f6144a8de94836c8c27b88f8a9bdcd88710
Reviewed-on: https://go-review.googlesource.com/c/tools/+/638096
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>