gopls/internal/util/astutil: NodeContains: workaround for ast.File
This CL causes NodeContains, when given a file, to report whether
the position lies anywhere within the file. In effect this is a
workaround for the unfortunate behavior of File.{Pos,End}, which
predates the addition of File.File{Start,End}, and it is consistent
with PathEnclosingInterval.

Change-Id: I90f8b6ea76fa78ba44401948eece303b4b2cb120
Reviewed-on: https://go-review.googlesource.com/c/tools/+/637738
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>