gopls/internal/golang: check for nil ReceiverNamed
This is a defensive check against a nil result from ReceiverNamed;

I'm unable to reproduce the panic though. It would have to be
hovering over a reference to a method whose receiver is
not a named type (or alias or pointer to named), such as
a slice type, an undeclared type, or somesuch.

Updates golang/go#70969

Change-Id: I1cc2bc3d311a88304a2af525d38c1143f4c513a1
Reviewed-on: https://go-review.googlesource.com/c/tools/+/638597
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>