internal/modindex: fix index directory for tests
The module cache index is written to a subdirectory
of os.UserCacheDir(). In non-test circumstances this
is the right choice, as it makes the index generally
available.

In tests this is the wrong choice, as it may leave test remnants behind.

This change sets the directory to a subdir  os.TempDir for tests.
Tests are free to further change modindex.IndexDir.

Change-Id: I046e0520adf3e3ffa71cfc8a0dfbdd060e41eb1a
Reviewed-on: https://go-review.googlesource.com/c/tools/+/639176
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>