gopls/internal/cache: remove overzealous bug.Report
defineView's doc comment confidently states that the only
possible failure is a canceled context...  yet the very
first thing it does is return an error if the folder path
is invalid---and it was ever thus since the comment was
added in CL 552315. D'oh!

Downgrading to a non-bug error.

Fixes golang/go#70909

Change-Id: Ic58199786816ad7de9f77d18a7d46795440bdf01
Reviewed-on: https://go-review.googlesource.com/c/tools/+/639435
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>