gopls/internal/cache/parsego: fix OOB panic in fixAST
The indices to the src[::] slice operation were not the
indices validated by the call to safetoken.Offsets.

I tried a couple of dozen incomplete expressions at EOF but
was unfortunately unable to trigger the crash. Still, this
is a principled fix.

Fixes golang/go#70745

Change-Id: Ia24c19e074517e8dd766cbf126d702f013392417
Reviewed-on: https://go-review.googlesource.com/c/tools/+/639397
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>