gopls/internal/analysis/unusedfunc: analyzer for unused funcs/methods
This CL defines a new gopls analyzer that reports unused functions
and methods using a local heuristic suitable for the analysis
framework, delivering some of the value of cmd/deadcode but with
the value of near real-time feedback and gopls integration.

Like unusedparams, it assumes that it is running within gopls'
analysis driver, which always chooses the "widest" package for
a given file. Without this assumption, the additional files
for an in-package test may invalidate the analyzer's findings.

Unfortunately a rather large number of marker tests define
throwaway functions called f that not trigger a diagnostic.
They have been updated to finesse the problem.

+ test, doc, relnote

Change-Id: I85ef593eee7a6940779ee27a2455d9090a3e8c7c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/639716
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>