gopls: prepare for mod cache index
This CL adds the infrastructure for using the module cache index when
satisfying missing imports. There is no change in behavior as it invokes
the existing imports.Source.

There is a new option importsSource whose value can be "goimports" to keep
the old behavior, "gopls" to use (in a future CL) the module cache
index, and "off" (for use by cider) to avoid looking in the file
system at all.

The index is kept in memory. Periodically the code checks to see if it
needs to be updated.

Change-Id: I61e0b5e224a6c26d75932417b26ecb9f432b460f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/640495
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>