go/ast/inspector: treat empty type filter like nil
The documentation for type filtering speaks only of the
length of the slice, but the logic checks for nil.
Fix the logic to match the documentation.

(This is an observable change but there is never any
reason to pass []ast.Node{}... as a type filter.)

Change-Id: Ifa8f72c0c2ff5a4c0fc2ee39e65641503a93c471
Reviewed-on: https://go-review.googlesource.com/c/tools/+/641436
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>