gopls/imports: use a module cache index
This CL switches to using an imports.Source that resolves missing
imports by first using the metadata graph (Snapshot.MetadataGraph())
to find missing imports from the workspace. Failing that it uses
the module cache index to find missing imports in the module cache.

In either case, it uses heuristics if there is more than one possibility.
For instance, it will prefer an alternative named in the go.mod file.
Also, it will tend to prefer a /v2.

The new behavior is controlled by an option 'importsSource'
which is (temporarily) set to get the old behavior, except in
some new tests.

Change-Id: I83a6f2bb12af94c9bc079a61306de3c9d1dc6ce4
Reviewed-on: https://go-review.googlesource.com/c/tools/+/641776
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Peter Weinberger <pjw@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>