internal/astutil/cursor: add Cursor.Edge method
This CL defines the Cursor.Edge method, which returns the
identity of the field in the parent node that refers to
the current node and, if it is a slice, the index.

For example, the CallExpr f(x, y) has three children, all
Idents: f has edge.CallExpr_Fun and index -1, and x and y
have edge.CallExpr_Args and indices 0 and 1 respectively.

The edge enumeration lives in the internal/astutil/edge
pacakge so that it can be shared by inspector and cursor.

The inspector records the edge and index during its walk,
and squirrels the information in the hitherto unused
event.parent field of each pop event. The edge and index
are packed into 32 bits, giving us an upper bound of 32M
elements in any []ast.Node, which seems like plenty.

Change-Id: Iad92d56629a119a9d69d486b2628ac9071f88bd3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/642478
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>