internal/telemetry/cmd/stacks: add GitHub client
Add a type for a GitHub client.
The type holds the auth token, and its methods are the only way
to access GitHub.

The first of two CLs that will make it possible to test more of this
program, by shunting changes to GitHub off to the side during testing.
(See the following CL for details.)

Change-Id: Ic487714fe75e19b016a132c0eeaaaf74d5c7cd42
Reviewed-on: https://go-review.googlesource.com/c/tools/+/643936
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>