gopls: remove go.sum files from integration tests
Explicit go.sum files in the txtar data for tests makes it
harder to modify the tests. This CL replaces the ones in the
gopls integration tests with calls to WriteGoSum in test setup.

And as part of modernization, a few 'interface{}'s have been
replaced with 'any's.

Change-Id: I951fa6d7b2ed780df68f4bc0d043f0738612da28
Reviewed-on: https://go-review.googlesource.com/c/tools/+/644335
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>