internal/refactor/inline: avoid crash when inlining empty function
When the call to an empty function is inside a go or defer statement,
the callStmt helper can return a nil ExprStmt. Previously, this would
result in a typed nil for res.old, leading to the crash of
golang/go#71486.

Fixes golang/go#71486

Change-Id: Ib956be2cf0d628567734f18b33e529999e081ab9
Reviewed-on: https://go-review.googlesource.com/c/tools/+/645816
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>