gopls/internal/analysis/modernize: mapsloop: fix two bugs
As with slices.Contains, there was a bug in the maps.Copy
modernizer resulting from not checking for implicit widening
conversions in m[k]=v; and another, from not checking that m
is a map. This CL fixes both.

Modernizers are hard. :(

Fixes golang/go#71313

Change-Id: Ie59aa9419868b3010435b7113bb5e67f0abbb4d3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/645876
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>