gopls/internal/analysis/modernize: strings.Split -> SplitSeq
This CL defines a modernizer for calls to strings.Split and
bytes.Split, that offers a fix to instead use go1.24's SplitSeq,
which avoids allocating an array. The fix is offered only
if Split is used as the operand of a range statement, either
directly, or indirectly via a variable whose sole use is the
range statement.

+ tests

Change-Id: I7c6c128d21ccf7f8b3c7745538177d2d162f62de
Reviewed-on: https://go-review.googlesource.com/c/tools/+/647438
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>