internal/analysisinternal: AddImport handles dot imports
If AddImport finds that an existing dot import suffices to refer to an
name, it returns that information by means of a first return value of
".", and does not add a new import.

For this to work, AddImport must know the name for which an import
is needed, so it can determine whether it is shadowed.

Change-Id: Ie4c9edf78fb89fc1b64f344517627173a253b999
Reviewed-on: https://go-review.googlesource.com/c/tools/+/647437
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>