gopls/internal/cache: downgrade bug reports for inconsistent metadata
As described in the lengthy comment included with this CL, it is
possible for gopls to encounter inconsistent metadata when it does not
observe all filesystem changes. This explains the frequent bug reports
of this nature that we have been seeing in telemetry. However, this is
not feasible to address without significant redesign of gopls' package
loading, and likely changes to the go command and go/packages
integration.

For now, we must downgrade the report.

Fixes golang/go#63822

Change-Id: I60d627eab00a33a1f68fdf9f2de9890bede33e73
Reviewed-on: https://go-review.googlesource.com/c/tools/+/647515
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Robert Findley <rfindley@google.com>