internal/analysisinternal: AddImport puts new import in a group
If AddImport needs to add a new import, and the file's first
declaration is a grouped import, then add it to that import.

This is one step towards a full implementation of the issue below,
and perhaps is good enough.

For golang/go#71647.

Change-Id: I8327b07c21c3efbd189c519e51c339b7aa4751d8
Reviewed-on: https://go-review.googlesource.com/c/tools/+/648136
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
Auto-Submit: Jonathan Amsterdam <jba@google.com>