internal/analysisinternal: disable AddImport test without go command
The AddImport test uses the default importer, which calls go list.
This fails in environments that can't call the go command, like
js/wasm.

Add a predicate to testenv that asserts the need for the default
importer, and call it from TestAddImport.

A subtlety: although this bug manifested itself only for the dot-import
cases, in fact all the test cases failed type checking on js/wasm for
this reason. But a successful type-check is not a precondition for
the test (see the new comment in TestAddImport). What caused the
particular test case to fail was a bad diff resulting from how the edit
was applied in the presence of that failure.

Fixes golang/go#71645.

Change-Id: Ib04afad108c323999bb67f329cf8d9cf329fead1
Reviewed-on: https://go-review.googlesource.com/c/tools/+/648580
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>