internal/analysis/gofix: simple type aliases
Offer to replace inlinable type aliases whose RHS is a named type.

Despite the amount of new code, there is little to test, because most
of it duplicates constants.

When there is a chain of inlinable type aliases, as in:

  type A = T
  type AA = A
  var v AA

we don't follow the chain to the end. Instead, the first set of fixes
produces:

  type A = T
  type AA = T
  var v A

That is, the replacements happen "in parallel." A second round
of fixes would rewrite

  var v A

to

  var v T

This case is rare enough that it's not worth doing better.

For golang/go#32816.

Change-Id: Ib6854d60b26a273b592a297cb9a650a31e094392
Reviewed-on: https://go-review.googlesource.com/c/tools/+/649055
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>