gopls/internal/analysis/gofix: inline most aliases
Support inlining an alias with an arbitrary right-hand side.

The type checker gives us almost everything we need to inline an alias;
the only thing missing is the bit that says that a //go:fix directive
was present. So the fact is an empty struct.

Skip aliases that mention arrays. The array length expression isn't
represented, and it may refer to other values, so inlining it would
incorrectly decouple the inlined expression from the original.

For golang/go#32816.

Change-Id: I2e5ff1bd69a0f88cd7cb396dee8d4b426988d1cc
Reviewed-on: https://go-review.googlesource.com/c/tools/+/650755
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>