gopls/internal/cache/parsego: add File.Cursor, and use it
This CL adds a Cursor to the parsego.File. Though the primary
motivation is convenience and flexibility, it is expected
to be an optimization: though it is computed eagerly,
it is retained in the parse cache, and is expected to pay
for itself very quickly by allowing us to replace many
whole-File ast.Inspect operations with more targeted traversals.

The CL replaces all ast.Inspect(file) operations with Cursor,
but there remain many more opportunities for using it in
narrower traversals, and in places that need to navigate to
siblings or ancestors.

Also, amend Cursor.FindPos to use the complete range of the
File, as CL 637738 recently did for astutil.NodeContains.

Also, various clean-ups to InlayHint:
- push the traversals down in InlayHint to avoid having
  to scan a slice for every single node we visit;
- simplify the function signature used for each hint
  algorithm.

Change-Id: I64d0c2cae75fd73a4b539ceb81ad9d6f7d80cfb8
Reviewed-on: https://go-review.googlesource.com/c/tools/+/650396
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Commit-Queue: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>