gopls/internal/golang: make singleFileFixer like fixer
In the past the singleFileFixer has two roles: to simplify
the signature for fixers of a single file, and to allow them
to be expressed only in terms of go/{ast,types} data types,
allowing fixers to be more loosely coupled to gopls. But that
latter role seems unimportant now, so this CL simplifies the
two functions to make them more alike.

Change-Id: I42ee9fd275e344fafee0b27b9861f8f599f89e3e
Reviewed-on: https://go-review.googlesource.com/c/tools/+/650641
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>