gopls/internal/golang: fix crash when hovering over implicit
After months of intermittent investigation, I was finally able to
reproduce the telemetry crash of golang/go#69362: if the operant of the
type switch is undefined, the selectedType will be nil, and therefore
logic may proceed to the point where it reaches a nil entry in
types.Info.Defs.

The fix is of course straightforward, now that we understand it: we
cannot rely on types.Info.Defs not containing nil entries. A follow-up
CL will introduce an analyzer to detect such problematic uses of the
go/types API.

Fixes golang/go#69362

Change-Id: I8f75c24710dbb2e78c79d8b9d721f45d9a040cd7
Reviewed-on: https://go-review.googlesource.com/c/tools/+/652015
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>