gopls/internal/util/asm: better assembly parsing
This CL adds a rudimentary parser for symbols in Go .s files.
It is a placeholder for a more principled implementation,
but it is sufficient to make Definition support control labels
(also in this CL) and for a cross-references index (future work).

+ test of Definition on control label
+ test of asm.Parse

Updates golang/go#71754

Change-Id: I2ff19b4ade130c051197d6b097a1a3dbcd95555a
Reviewed-on: https://go-review.googlesource.com/c/tools/+/654335
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>