internal/analysis: add function to delete a statement
In gopls, for unused variables and for modernizers, there is a need
to provide edits to delete a specific statement. To avoid duplication
these will use this new DeleteStatemnt. This CL, the first
of a series, is to install the new function and its signature,
and provide tests.

Immediatley following CLs will use the new function to
replace existing code.

Change-Id: I1213cfaf14e66eaa9a111b11094a536bb869b298
Reviewed-on: https://go-review.googlesource.com/c/tools/+/655295
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>