gopls/internal/golang: unify tracks type params
Enhance unify to take type params into account.

Unify(x, y) will return false if there is no assignment to the type
parameters of x and y that will make x identical to y.

(Except that type param constraints and interface literals are not
handled; unify "fails open" for these, returning true when the right
answer might be false. That is the same behavior as previously.)

The API supports initial bindings for type params, in order to handle
types like C in instantiations of F:

    func F[T any]() {
      type C *T
    }

The implementation matches that in internal/util/fingerprint.go,
except that it works on actual types.Type values instead of the
reconstituted fingerprint types.

This CL tests the return value when there are no initial bindings.
Subsequent CLs will test the final values of bindings, with and
without initial values.

Change-Id: I2402f818b86a5ccac874491e0801bb503b449cd6
Reviewed-on: https://go-review.googlesource.com/c/tools/+/657076
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>