gopls/internal/analysis/modernize: preserves comments in minmax
This CL changes the original deletion (from after rh0
to the end of the if stmt) into deletion from the start of assignment to
the end of if stmt), add all comments between them and create a new
assigment as-is.

This change preserves all comments inside if stmt and the comments
after the line of assignment and before if stmt, causing comments B,C,D
to be preserved and put on the top of min/max function call after fix.

- source:

	lhs0 = rhs0 // A
	// B
	if rhs0 < b {
	    // C
	    lhs0 = b
	    // D
	}
- fixed:
	// A
	// B
	// C
	// D
	lhs0 = max(rhs0,b)

Fixes golang/go#72727

Change-Id: I7c193711aac5834ebb0d5e8ae22c26ae7990c34f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/656655
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>