go/analysis/passes/printf: disallow %q of int
This CL causes the printf checker to report a diagnostic if an
arbitrary integer other than byte or rune is formatted with %q.
This is a common mistake. (However, uses of %c and %U with such
integers are rarely mistaken.)

The check only applies to files using go1.26 or later.

+ test with go1.25 and go1.26

Fixes golang/go#72850

Change-Id: Ic09a1d726e99608e3067f5e1a37bacc3a835724f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/657957
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>