gopls/internal/cache: improve build constraint trimming
Generalize trimContentForPortMatch to handle +build directives.

It assumed only go:build directives, but the +build variety is still
valid, and in fact there is a file in the Go build on my local
Google-internal machine that has them.

This fixes a test that was failing for me because of that file.

Change-Id: I534a18ef6e66575d242406e7b81c32055e3c8ace
Reviewed-on: https://go-review.googlesource.com/c/tools/+/658195
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>