internal/astutil/cursor: unsplit Parent{Edge,Index} -> ParentEdge
This CL is a partial revert of CL 659575, which was merged by
mistake. During review, we agreed that the original form
was more logical (if sometimes inconvenient), but that
we liked the new name, and the new ChildAt function.

Change-Id: Ib1d4e0dfa8cefdb944eb1394be946006e8285390
Reviewed-on: https://go-review.googlesource.com/c/tools/+/659635
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>