gopls/internal/cmd: suppress TestImplementation on go1.23
The types.CheckExpr data race (#71817), fixed at master and
backported to go1.24, is not backported to go1.23. The race
itself is benign, but it causes this one test to flake.
So we suppress it on go1.23.

Fixes golang/go#72082

Change-Id: I64731adc50137aefbbecc0b7a47a41036d831eab
Reviewed-on: https://go-review.googlesource.com/c/tools/+/661176
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Griesemer <gri@google.com>