gopls/internal/analysis/modernize: waitgroup: use index.Calls
This CL changes the new waitgroup modernizer to use index.Calls
to enumerate the calls to WaitGroup.Add directly, instead of
searching for all "go" statements. This is an optimization,
though only a minor one because go statements are already
sufficiently rare that it doesn't matter. The real purpose
of the change is to try to establish the form that we wish
other modernizers (which may search for more numerous nodes)
to follow.

Also, remove check for uses of WaitGroup type itself.
The modernizer's pattern doesn't depend on the type.

Change-Id: Ie87a33b08b71764ced13f204e5e0b6e0ed35d58f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/662276
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Commit-Queue: Alan Donovan <adonovan@google.com>