gopls/internal/server: workspace: skip adding already added file
This CL adds a check to prevent re-adding a folder to the workspace
if it has already been added. It resolves an issue where the 'loading
packages...' status bar would get stuck in the editor.

The problem occurs when the editor sends an initialization request with
a rootURI but no workspaceFolders, followed by a DidChangeWorkspaceFolders
request for the same folder. Re-adding the same folder is unnecessary and
causes the status bar to hang.

Additionally, this change cleans the provided URI to handle cases where
the client appends a trailing slash to the same folder.

An alternative approach is to reload the added folder again and end the
process of loading package as expected so the status bar could end soon,
which is unneccessary because there is no change in given file.

Fixes golang/go#71967

Change-Id: Ib28c327a5a85f1fdd54e4facedf97d500b909fb2
Reviewed-on: https://go-review.googlesource.com/c/tools/+/663295
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>