internal/stdlib: find api directory more robustly
stdlib contains precomputed information about exported symbols in the
standard library. The old code looked in runtime.GOROOT()/api for
this information. This works on many systems, but not always.
The new code runs go env, and looks first in GOROOT/api, and failing
that, in GOPATH/api.

(The situation seems to depend on mysterious details of the history of
the user's Go environment.)

This CL is on the path to adding function snippets for the standard
library, for the new unimported completions code.

Change-Id: If2853a3f26edf9d0f1603563104592b95eb466a1
Reviewed-on: https://go-review.googlesource.com/c/tools/+/664995
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
Commit-Queue: Alan Donovan <adonovan@google.com>