gopls/internal/golang: Rename all receivers only at declaration
This change modifies the behavior of the new "rename receiver"
operation, which affects the receivers of all methods of the
same type: now, the broader renaming occurs only when the
renaming is requested at the declaration of the receiver.
Renaming a use of a receiver affects only that variable.

+ test, doc, relnote

Fixes golang/go#41892

Change-Id: I54c771f51da90b8b2a25801435e81875639ab459
Reviewed-on: https://go-review.googlesource.com/c/tools/+/666455
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>