gopls/internal/golang/completion: new code for unimported completions
The unimported completion code is invoked when the user is looking
for a package-level symbol in a package that has not been imported
into the current file. The code looks for matching symbols in a number
of places.

If the user has typed foo.lw (with the cursor just after the 'w'), then
a matching symbol will be exported from some package 'foo', with
'l', 'w', a subsequence of its name (converted to lower case). That is,
the typed symbol name to the left of the cursor is treated as a pattern.

The code looks first for a match in the other files of the current
package. Failing that it looks in the standard library, using the
stdlib.PackageSymbols built into gopls. Failing that it looks in
the rest of the workspace, and failing that it looks in the module
cache.

Test have been changed to use the new code. (An alternative would be
to duplicates the relevant tests, one version using the old code and
the other using the new code.)

Change-Id: Ie04e5571ce09f637fc3913f8e2c51aa65704844f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/667576
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>