go/ssa: clear Function.subst after building bodyless function
The subst field is unlike the other ones because it is set
non-nil at construction, not during building, so it is not
sufficient to clear it in finishBody, which is not called
for bodyless functions. This change causes it to be cleared
in buildParamsOnly too, and documents the coupling between
the two operations.

+ test that the sanity check no longer fails

Fixes golang/go#73594

Change-Id: I72b471bf1596493f12f6cd9f3ca59a055834aad5
Reviewed-on: https://go-review.googlesource.com/c/tools/+/669955
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Elias Naur <mail@eliasnaur.com>
Commit-Queue: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>