gopls/internal/golang: fix crash in definition of invalid "continue label"
Telemetry reported an out-of-bounds panic for a definition request
on the "continue" keyword in a "continue label" statement.
A label with an invalid location would not be found on the path
enclosing interval, causing this error. This is illustrated by
the following example:
```
func F() {
    label:
	for i := range 10 {
	}
	for i := range 10 {
		continue label
	}
}
```

Add a bounds check and return nil if the label is not found.

Update the marker test for @def to accept variadic arguments.
If no argument for dst is specified, assert that there is no
definition returned - the value will be an empty protocol.Location.
@def now also accepts a named argument for err.

Add a new test wrapper GoToDefinitions that returns all result
definitions.

Fixes golang/go#73797

Change-Id: I266079251eb79646a7906b11286d8b20ceb966ac
Reviewed-on: https://go-review.googlesource.com/c/tools/+/674497
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>