gopls/internal/analysis/recursiveiter: report inefficient recursion
This CL adds a new gopls analyzer, recursiveiter, that reports
inefficient recursive uses of range-over-func in the implementation
of an iterator (iter.Seq or iter.Seq2). The issue is rather subtle
and the problem hard to see, but it is quite easy to detect
(though not fix) analytically.

Here's an example bug from the module mirror corpus:
https://go-mod-viewer.appspot.com/golang.org/x/arch@v0.17.0/internal/unify/yaml_test.go#L84
(I don't expect there will be many yet since iterators are new,
but the check is cheap.)

Also: report an error if a marker test file has duplicate
sections. Thanks to Rob for pointing out my error.

+ test, doc, relnote

Change-Id: If9e625bf321aa5fa3a2b1f2ab0a65f31d445be70
Reviewed-on: https://go-review.googlesource.com/c/tools/+/674438
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>