internal/tokeninternal: optimize AddExistingFiles
A sequence of calls to AddExistingFiles typically causes
a large FileSet to grow in small amounts, resulting in
a linear number of calls to O(n log n) sort operations.

This consumed 12% of CPU on gopls DidChange/kubernetes benchmark
on a file with a large upward dependency graph: the type-
checking batch is approximately 25,000 files, and the FileSet
grows to this size across thousands of calls.
This fraction grows with the size of the project and appears
to be the cause of field reports of poor performance.

More efficient is to sort only the handful of new files,
then merge with the existing ones, which are already
sorted. In-place merging with constant space is too
fiddly, so we allocate additional 2x space in the FileSet
array and use the top half as the second buffer.

This change improves the benchmark from CL 675875 by
almost 10x. The more principled asymptotic improvement
in CL 675736 will have to wait for go1.25.

Apple M1
                                    │    old.txt    │              new.txt               │
                                    │    sec/op     │   sec/op     vs base               │
FileSet_AddExistingFiles/sequence-8   387.91m ± 17%   53.28m ± 3%  -86.27% (p=0.000 n=9)

Change-Id: I04c3b76ff2df50b6413206f057455366ca71caef
Reviewed-on: https://go-review.googlesource.com/c/tools/+/675535
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>